Separate welcome and tutorial.
