+++ src/App.tsx
@@ -3,6 +3,7 @@ import { Redirect, Route } from "react-router-dom";
 import { IonApp, IonRouterOutlet, IonLoading } from "@ionic/react";
 import { IonReactRouter } from "@ionic/react-router";
 import Quiz from "./pages/Quiz";
+import Tutorial from "./pages/Tutorial";
 import Question from "./pages/Question";
 import Incorrect from "./pages/Incorrect";
 import Correct from "./pages/Correct";
@@ -68,6 +69,11 @@ export const App: React.FC = () => {
             exact
             path="/quiz"
             render={props => <Quiz {...props} quiz={quiz} />}
+          />
+          <Route
+            exact
+            path="/quiz/tutorial"
+            render={props => <Tutorial {...props} quiz={quiz} />}
           />
           <Route
             path="/quiz/question"
